finish tests and split to separate testing file
